Successful trouble shooting: FSI runs without diverging.
Yet to confirm whether agreement with benchmark is achieved.
Fixed two significant bugs. First is shifting the assembly of the fluid variational form to within each time step so that it sees the updated mesh. Second is introducing a bounding box tree allows mesh velcoity bc on fluid to be implemented consistently.

Need aligngment with benchmark and a faster solver.